EMSUSD-1521 refactor AE template
- Move each custom control to its own file.
- Rename custom control and their file to follow the same pattern.
- Move the creator function to static function in the corresponding custom control.
- Move some helper functions to their corresponding custom control.
- Rename some parameters to have clearer name (attrNames instead of controls and attrName instead of c).
- Document how the AE template code works.
