+++ lib/mayaUsd/resources/ae/usdschemabase/attributeCustomControl.py
