+++ lib/mayaUsd/resources/ae/usdschemabase/materialCustomControl.py
